+++ orders/graphql/resolvers/order_resolvers.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from orders.exception import custom_exceptions
 from orders.graphql.types.error_types import OrderNotFound
 from orders.graphql.types.types import OrderType, OrdersType
@@ -16,10 +18,10 @@ def map_order_response(order_dto) -> OrderType:
         restaurant_id=str(order_dto.restaurant_id),
         promo_code_id=order_dto.promo_code_id,
         status=order_dto.status.value,
-        items_total=float(order_dto.items_total),
-        delivery_fee=float(order_dto.delivery_fee),
-        tax_fee=float(order_dto.tax_fee),
-        final_amount=float(order_dto.final_amount),
+        items_total=Decimal(order_dto.items_total),
+        delivery_fee=Decimal(order_dto.delivery_fee),
+        tax_fee=Decimal(order_dto.tax_fee),
+        final_amount=Decimal(order_dto.final_amount),
         address_id=order_dto.address_id,
         placed_at=order_dto.placed_at,
     )
